requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
audioread==3.0.1
blinker==1.9.0
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
crispy-bootstrap5==2025.6
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
Django==5.2.11
django-admin-rangefilter==0.13.5
django-bootstrap5==25.2
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
dnspython==2.4.2
et_xmlfile==2.0.0
Flask==3.1.2
fonttools==4.60.1
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
googlemaps==4.10.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
moviepy==2.2.1
msgpack==1.1.2
mysql-connector-python==9.3.0
mysqlclient==2.2.7
noisereduce==3.0.3
numba==0.62.1
numpy==2.3.1
oauthlib==3.3.1
openai==2.15.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pillow==10.4.0
platformdirs==4.5.0
pooch==1.8.2
proglog==0.1.12
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
PyJWT==2.11.0
pymongo==4.6.0
PyMySQL==1.1.1
pyparsing==3.2.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
requests-file==3.0.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
spyne==2.14.0
sqlparse==0.5.5
threadpoolctl==3.6.0
tkhtmlview==0.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.5
zeep==4.3.2
